+++ src/features/auth/types.ts
@@ -1,6 +1,7 @@
-export type LocalUser = {
+export type User = {
   firstName: string | null;
   role: string | null;
   isAuth: boolean;
   token: string | null;
+  theme: 'dark' | 'light';
 };